move mcp stuff into a subfolder
