+++ react-sdk/src/index.ts
@@ -47,4 +47,3 @@ export {
   type TamboThreadMessage,
 } from "./model/generate-component-response";
 export { type TamboThread } from "./model/tambo-thread";
-export { TamboMcpProvider } from "./providers/tambo-mcp-provider";
