+++ react-sdk/src/mcp/index.ts
@@ -1,0 +1,1 @@
+export { TamboMcpProvider } from "./tambo-mcp-provider";
